cleanup forms.py
