+++ indicators/widgets.py
@@ -1,3 +1,5 @@
+from django import forms
+from functools import partial
 from django.forms.widgets import Select
 
 
@@ -15,3 +17,13 @@ class DataAttributesSelect(Select):
         for data_attr, values in self.data.iteritems():
             option['attrs'][data_attr] = values[option['value']]
         return option
+
+
+class DatePicker(forms.DateInput):
+    """
+    Use in form to create a Jquery datepicker element
+    Usage:
+        self.fields['some_date_field'].widget = DatePicker.DateInput()
+    """
+    template_name = 'datepicker.html'
+    DateInput = partial(forms.DateInput, {'class': 'datepicker'})
